trim files in internal/rbac
