--- internal/rbac/rbac.go
@@ -1,1 +1,0 @@
-package rbac
